Early stages of bid implementation. (see BidDBServlet.java), read body
So far, I am trying to figure out how to get the auctionID string
from auction.jsp to my Java servlet, BidDBServlet.java.

I see how the auction_id is retrieved in line 39 of auction.jsp,
but that string cannot be retrieved after leaving that page
(it wasn't user input, as per a form)

(if you look in BidDBServlet.java, I try to call
request.getParameter("auction_id"); to get the auctionIDStr, but this
only returns null -- and thus the page redirect fails to work)

The page redirect is on line 97 of BidDBServlet.java,
and without getting the auctionIDStr -- you'll end up with this:
"auction.jsp?auction_id=NULL"

instead of the auction_id being in place of NULL.

If I have some idea as to how navigating through auctions will go,
either by the User's homepage or by search --
I may be able to use a SQL query to match the hyperlink string
(which would be the description, or title of the auction)
with one of that within the table -- and thus I have a way
of getting the auction_id.

I will need the auction_id to redirect the user back
to the auction page once an attempt to make a bid is complete.

Assuming a bid made by the user is greater than or equal to
the minimum bid plus the increment amount
(userBid >= minimumBid + increment), that bid will be allowed to go
through and will be logged into the database.