Bidding still needs more work (see body) Must solve bidding issue.
Issue #1:
I am able to insert a bid tuple into the database for my own bids,
but cannot insert a bid tuple into the database that are of other
user's auctions -- which is sort of the point!

For now,
when a bid is made -- it will at least be able to redirect
you back to the same auction page from which you came.

Issue #2:
I also need to show an error to the user
if the bid amount specified is too low.

Issue #3:
If I can get the above working, then
I can write the update SQL statement to update
the minimumBid amount for the current auction.

Issue #4:
Also need to check if a bid is being made within
the start time and end time of the auction
before even checking the bid amount --
but I need to know how to show error messages
to the user to confirm this.